indicadores/requirements.txt
dnspython==2.0.0
pymongo==3.11.0
python-decouple==3.3
requests==2.24.0
